@samnpowell,rdimaggio
Good job! A lot to do in this homework. Did you get an error on the SOQL practice when you ran in execute anon?
When I ran  WeekFiveHomework.soqlPractice(); I got an error- see my comments on line 15. Apex/SOQL is very particular about having all the fields listed in your query that you want to use. So if you want to use contact LastName in your loop, or in any logic in this class, you will have to include the LastName field in your SELECT query. This will get all those LastNames out of the database and put them into the collection, the list <contact> contacts, that you have created. Then you can use that field in your loop etc.
I also added a counter variable on line 54 for variety. I cant remember if that is what line49 is asking for as I accidentally overwrote the original homework template with someone's homework!! (silly me). But your original debug  with +i also worked fine! Anyhow, nice work and great catchup. Week 6 homework only has a wee edit and 2 lines of code to add, so if you haven't done it, know that you are nearly there. See contactTriggerHandler onAfterUpdate for a template for week 6 homework if you get stuck (or just start there- starting with a working template is what I do a lot of the time!!)!